Move nav item create from drawer to body
